GUI setup and main menu
This code is the base of my GUI for all the frames, it's set up for things to be added easily in the future. I also created the first  frame which is the MainMenu Frame, the buttons on the main menu frame don't work bc the method for the command isn't created yet. It's also missing the "would you like to choose from" and the image.