New Main Game Class
I created a new main game class but don't have anything that runs and constructs the stuff in the class nor do I have anything to switch between the frames. I made a frame, a canvas within that frame to create another frame within that canvas to have the boxes for the notes and then It's a canvas so I can add a scroll wheel because I have a lot of notes.