New FrameManager Class
As I decided to create different classes for the different frames I need to do an entire overhaul on how my GUI system worked. To do this I created another class called Frame Manager. This class is the parent class for all of my "subclasses" which are the different frames. like how I did it before I created a frame to hold all of the other frames and raised them using frame.tkraise. In this class as this will be the class that is being "run" and the parent of the other classes I made it load the scent_data from a JSON file containing all of the scents. Since I added this I needed to reformat the rest of the game. the MainMenu Class is just a copy paste of it before with constructing also from the parent, maingame is the same and pretty much unchanged. and the paleteeSelector as that will use different methods now I have just # it out and will work on it later.