Created the grid and added the scents
I created the create_scent_boxes_grid method for the maingame which adds all the boxes containg the scents and their attributes into the left side "scents_grid_frame" that I created. I have to fix the reading from the file and there is an issue with the colums, I will fix.