Fixed Scent Boxes
I did a bunch of small changes like typo fixes, resizing, etc.
I fixed the typo that created an error, I fixed the resizing of the canvas and the scroll wheel, I created a frame for the back and checkout button and I also added a checkout button to go to the checkout. I fixed my selected_scent_var thing by adding it to a list because otherwise it just overwritted the previous one wiht the for loop. I also fixed the sizing for the scent_boxes. I also added relief="solid" to make the border for each box show becasue it wasn't previously.