Made the Checkout Frame
I created the checkout frame and added labels, entry boxes to prompt the user to enter a name. I also binded the return key to a method that would show the final perfume fragrance profile, the name of the perfume, the chosen scents etc. I haven't done the combined totals on the MainGame Frame yet but once I have I will pass those values over to be added here.  I have hashed out the image stuff as the Image stuff was giving me problems in the MainMenu, once I have found how to get the image to work there I will do the same thing for the checkout. Next Steps are to do the combined totals stuff and to do the checkout and MainMenu image.