Working totals
I made the totals actually work. I created the totals labels on the side showing all the attributes set at the beginning. I then created a method that would set the totals to 0 then for every scent in the list that saves the choice selected using the combo box that it would get that data and add it to a running total. I binded this method to the event of selecting the combobox so that it automatically updates when the user picks something in the combobox.