Added Totals to the End, Added Comments
I added the scent totals into the show_final_scent_name method so when the user presses return it shows the name AND also the scent totals, these where already saved in scent_totals in the framemanage and using super().__init__(parent) I could take that value and use it here. I just took it and put it on a seperate line. I also added comments for the things in my code