New Messagebox
I added a messagebox when the user clicks on the checkout button, confirming if the choices they made are the ones they want. The way my code is structured and the way the totals load I could've had a button but I felt as if a back button in the checkout frame would be very odd. I also fixed the if_scent_name ==True thing mentioned in the comments of the last commit.