Add Button, Reset Button, Reset Method
I added an add button for every scent box created in the for loop, it has no command as of now, I also added a reset button and a reset method, I structured how the reset button will work but I still need to create the necessary variables for that to work so its ### for now. I also removed the list containing the combobox stuff. I also added a selected_scents list to put the selected scents into, currently this whole section doesn't work as im doing some here and there but the beginning for it is there.