Add Button, Reset Button Method
Everything that I goes in the scent boxes is the same as before apart from, now I just need to add a button. As the creation of the boxes are  in a for loop I could just add the button and the .grid for the button to the loop and it'll apply to them all (which is what I did) . I also created the reset_selections method that is already tied to the reset button, it just takes the select scents list and configures it to show the label as "scent: none" (show its empty). select_scent is a method that's empty right so the adding, the updating and the reset doesn't currently work but once I add that method it should all work. (doesn't work becasue how can you update after adding if you havent added).