Working Select_scent method.
add self.selected_scent_names = [] to frame manager for access in checkout class. I also created the select_scent method a which checks if the selected_scents list is not more than 3, then if its not then it appends the scent_name to the selected_scents list , then runs the update totals and display methods so the user can see it update in real time. I also added a messagebox to the else just to let them know that the limit is reached. added a self and a 2 instead of one to fix mistake elsewhere in code.