New BaseFrame class, New Palette Button setup
I added a baseframe class which is now the parent to the main menu and the palette selector frame. In both of those classes the configure for the rows and columns were the same and had the same weight so by adding this I need to do as super init like how I used to with the framemanager, all I need to do is pass the number of rows and columns and it’ll set it up with weight = 1 for however many there are. I also added a create_styled_button and palette button to the main menu and palette selector to simply creation as they are all the same. I also removed the names for the palette buttons. Since I just use an else statement start main game method because I don’t mention the palette name I can remove it, since it's redundant.