Added Validation
Checks if inserted name is alphabet, added validation for if json error if there is then closes, added validation to make sure that at least one scent is chosen, added a messagebox check / confirmation for the reset button, checks if scent_notes and palettes are in the data. Checks if the scent in the palette is in the act data if not then it’s missing it’s attributes and the code cannot function or calculate properly.